chore: meteor collection implementation refactoring
